refactor(oh7): move SqliteLog out of core into the storage-backed bin (#252)
