+++ src/cli/run_es_record.rs
@@ -333,8 +333,9 @@ pub fn record_ring_es_into(
 /// is malformed (e.g. no `RunStarted`) or storage fails.
 #[cfg(feature = "storage")]
 pub fn project_run(db: &crate::storage::Database, run_id: &str) -> anyhow::Result<()> {
-    use crate::core::orchestration::es::log::{EventLog, SqliteLog};
+    use crate::core::orchestration::es::log::EventLog;
     use crate::core::orchestration::es::state::fold;
+    use crate::storage::es_log::SqliteLog;
 
     // 1. Read the event log for this run.
     let log = SqliteLog::new(db.clone());
@@ -804,7 +805,7 @@ mod storage_tests {
         let db = init_embedded().unwrap();
 
         // Persist a minimal blackboard log via SqliteLog.
-        let mut log = crate::core::orchestration::es::log::SqliteLog::new(db.clone());
+        let mut log = crate::storage::es_log::SqliteLog::new(db.clone());
         for e in sample_blackboard_events("run-x") {
             use crate::core::orchestration::es::log::EventLog;
             log.append("run-x", &e).unwrap();
